Move more orphans
